Optimize (#18)
* added stopwatch

* added range thing

* changed float an moved stuff to optimize

* cleanup

* Organize

* Test framework

* Add thread safe list

* fix

+++ server/src/main/java/no/fractal/database/GateQueries.java
@@ -22,7 +22,7 @@ public class GateQueries extends PsqlDb {
 
         sqlQuery(query, resultSet -> {
             ret.add(new TensorData(
-                    (BigDecimal[]) resultSet.getArray("face_vec").getArray(),
+                    (double[]) resultSet.getArray("face_vec").getArray(),
                     UUID.fromString(resultSet.getString("user_id"))
             ));
         });
